physio_trial/requirements.txt
customtkinter>=5.2.2
CTkMessagebox>=2.5
pandas>=2.1.0
openpyxl>=3.1.2
numpy>=1.26.0
python-dateutil>=2.8.2
SQLAlchemy>=2.0.30
psycopg[binary]>=3.1.18
python-dotenv>=1.0.1
inject>=5.3.0
